+++ src/components/Footer.tsx
@@ -1,0 +1,4 @@
+"use client";
+export default function Footer(){
+  return <footer className="border-t border-neutral-200 bg-white text-neutral-500 text-xs text-center py-4 px-4">Created by — Mohd. Izhan</footer>;
+}
